Refactor HTTP methods and import declaration in 'BookController'
Refactored the HTTP methods and import declaration in 'BookController'. Replaced multiple import statements with a single one that uses '*', simplifying the declaration. Additionally, changed '@GetMapping' to '@PostMapping' to better reflect the intended use of the API endpoint '/allWithSpec'. These changes enhance code readability and architectural ethics of the CRUD operations performed in the API.